resume-insight-backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
PyPDF2
requests
